pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devops-control-plane"
version = "0.1.0"
description = "A general-purpose project control plane for DevOps automation"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "jfortin" },
]

dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
web = [
    "flask>=3.0",
    "cryptography>=41.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
    "types-click>=7.0",
    # Include web deps in dev for full testing
    "flask>=3.0",
    "cryptography>=41.0",
]

[project.scripts]
controlplane = "src.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ── Pytest ───────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external tools",
]

# ── Mypy ─────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false          # gradual typing — tighten later
check_untyped_defs = true
ignore_missing_imports = true          # third-party stubs may be missing

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
